[DM-37408] Add SODA Datalink Notebook
This is a basic notebook on how to use the datalink service to get the URL
to the cutout service and use it for creating cutouts.  We also download
the full image to have something to compare against to ensure that the
cutout is an actual cutout of the larger image.